basic funcs ready

+++ my_package/api_v1/schemas/review.py
@@ -7,7 +7,7 @@ from typing import Optional
 class ReviewBase(BaseModel):
     user_id: int
     place_id: int
-    rating: int = Field(default=1, ge=1, le=5, description="Рейтинг от 1 до 5")
+    rating: float = Field(default=1, ge=1, le=5, description="Рейтинг от 1 до 5")
     comment: Optional[str] = Field(default=None, max_length=500, description="Комментарий")
